+++ tests/Core/NumArrayTest.php
@@ -46,4 +46,21 @@ class NumArrayTest extends \PHPUnit_Framework_TestCase
         );
         $this->assertSame(6, $numArray->getSize());
     }
+
+    public function testGetNDim4()
+    {
+        $numArray = new NumArray([8, 9, 3, 6]);
+        $this->assertSame(1, $numArray->getNDim());
+    }
+
+    public function testGetNDim2x3()
+    {
+        $numArray = new NumArray(
+            [
+                [8, 6, 4],
+                [3, 8, 0]
+            ]
+        );
+        $this->assertSame(2, $numArray->getNDim());
+    }
 }