+++ CHANGELOG.md
@@ -11,3 +11,4 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add function `NumPHP\Core\NumArray::getData`
 - Add function `NumPHP\Core\NumArray::getShape`
 - Add function `NumPHP\Core\NumArray::getSize`
+- Add function `NumPHP\Core\NumArray::getNDim`